# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Totem",
        "description": "Documentação da API Totem com autenticação por token de empresa e token de dispositivo (TV).",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/departamentos": {
            "get": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Lista departamentos da empresa autenticada",
                "description": "Display a listing of the resource.",
                "operationId": "8f67cca908527786195d5efd02911ab4",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "CompanyBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Cria departamento",
                "operationId": "c288cfc72036fe95d515c7c0ef3df4f7",
                "responses": {
                    "201": {
                        "description": "Criado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "CompanyBearer": []
                    }
                ]
            }
        },
...
        "/api/produtos": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Lista produtos da empresa autenticada",
                "operationId": "ff4eeb91e587a0df77631ccd714d9344",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "CompanyBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Cria um produto",
                "operationId": "0f94cb8e97d212f1b83331b15ecc3862",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProdutoPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "CompanyBearer": []
                    }
                ]
            }
        },
        "/api/produtos/{produto}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Busca produto",
                "operationId": "e35c815ea8ee79452fb9210509f0e96d",
                "parameters": [
                    {
                        "name": "produto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "CompanyBearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualiza um produto",
                "operationId": "ab0172a44c47da4da79063dc385ea0dd",
                "parameters": [
                    {
                        "name": "produto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProdutoPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "CompanyBearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Remove um produto",
                "operationId": "21de4bdcad844fc813a2fed0f0c5ef1e",
                "parameters": [
                    {
                        "name": "produto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removido"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "CompanyBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ProdutoPayload": {
                "required": [
                    "NOME",
                    "PRECO",
                    "departamento_id",
                    "grupo_id"
                ],
                "properties": {
                    "CODIGO": {
                        "type": "string",
                        "maxLength": 14
                    },
                    "NOME": {
                        "type": "string"
                    },
                    "PRECO": {
                        "type": "number",
                        "format": "float"
                    },
                    "OFERTA": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "IMG": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "departamento_id": {
                        "type": "integer"
                    },
                    "grupo_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "CompanyBearer": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            },
            "DeviceBearer": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}